Additional logic and messages for existing and additional modules
- Added logic to detect SpamAssassin's preferred Geo::IP, using
  IP::Country::Fast as a fall-back option, and logic to inform user if one,
  both, or neither are found.

- Added basic logic to detect Debian/Ubuntu/apt-deb distributions.
  Installing the "razor" package will always resolve a dependency on an SHA1
  implementation.

NOTES:  As of Wheezy, the dependency on Digest::SHA1 has been dropped, along
with the package "libdigest-sha1-perl" itself, in favor of Digest::SHA,
which is bundled in the core distribution of Perl itself, the package
"perl".  The base source package for Razor2 has a debian/ tree which patches
the code to use Digest::SHA which is bundled in the core Perl distribution
package, "perl".  Squeeze and older require it but install the package
"libdigest-sha1-perl".  In both cases, the dependency is resolved when the
"razor" package is installed.